Implement basic subscriber importing from csv
To handle tsv or other similar files, will need to detect the delimiter
and handle the specific cases in parseSubscriberList and
countColumnsInFile.

The former (parseSubscriberList) relies on .join(',') to form the headers (fields in csv). A non-join alternative might
be to find a way to specify headers and pass a specific headers obj to
the csv parser (using Papaparse).

The latter (countColumnsInFile) just does a dumb .split('\n').split(',').length to count the columns, so the argument to the 2nd split will need to be modified for tsv/csv/whatever.

Need to test how well these operations work on huge files too!