Fix: stop new subscribers from always being rejected as duplicates
Finding duplicate subscribers using:
    SubscriberModel.find({email: subscriber.email}, (err, doc) => ...
and then detecting them using:
    if (doc) {... reject response}
doesn't work because doc is always an array, and even an empty
array is truthy, so the rejection function is always called.

This is fixed by using doc.length as a condition instead.
i.e.
    if (doc.length) {... reject response}

* Might be better to use .findOne (instead of .find) to simplify
this, because we only care if there is at least one subscriber.