Fix: prevent campaign sending from ending prematurely
Occasionally, campaign sending stopped prematurely without a crash and
without a success/err message.

This was because .findOne in getInitialListSubscriberId() did not always
return the id of the first subscriber in a list, due to wonky ordering
in the Postgres database. As a result, email sending would skip a number
of subscribers in a list (+ recipients were not predictable because
imported csv rows do not correspond or map directly to db rows).

Fix by sorting by ascending id first so that ListSubscriberIndexPointer
intially refers to the first ListSubscriber in a List.

There should not be a significant cost to this sorting, since
getInitialListSubscriberId() is only called once at the
beginning of campaign sending.